connector_evals/pyproject.toml
[project]
name = "connector-evals"
version = "0.1.0"
description = "Pydantic-evals based evaluation framework for testing connector quality"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Elastic-2.0"}
authors = [
    {name = "Airbyte", email = "contact@airbyte.io"}
]
keywords = ["airbyte", "connector", "evals", "pydantic-ai", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pydantic-ai>=1.0.8",
    "pydantic-evals>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "httpx>=0.24.0",
    "jsonref>=1.1.0",
    "jsonpath-ng>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["connector_evals"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "evals: marks tests as evaluation tests (may be slow)",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
